Version 2.1
Some correction regarding the use of attributes. Correct use of the ' is a bit tricky. Also improved the search for begin of word which was quite buggy.

I have created a complete description on how the omni completion works on wikibooks [1] if you like to know how it all works or prehaps like to create omni completion for another lanugage.

[1] http://en.wikibooks.org/wiki/Learning_vi:Vim#Onmi_Completion